Year-3/Hand Gestures Studio/requirements.txt
customtkinter
numpy==1.26.1
pandas==2.2.3
Pillow==10.1.0
opencv-python
matplotlib==3.10.0rc1
seaborn==0.13.0
tensorflow==2.18.0
mediapipe==0.10.13
scikit-learn==1.4.1.post1
h5py==3.11.0